+++ stts_platform_interface/lib/stts_platform_interface.dart
@@ -1,4 +1,5 @@
 export 'src/stt/model/stt_recognition.dart';
+export 'src/stt/model/stt_recognition_options.dart';
 export 'src/stt/model/stt_state.dart';
 export 'src/stt/stt_platform_interface.dart';
 export 'src/tts/model/tts_state.dart';