+++ stts/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.9.1
 * feat(Android/STT): Add `muteSystemSounds` to mute default system beep sounds.
+* feat(STT): Add recognition options.
 
 ## 0.9.0