+++ stts_web/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.9.2
+* feat(STT): Add recognition options.
+
 ## 0.9.1
 * fix: pub.dev score.